requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==44.0.0
cssselect2==0.7.0
dacktool==0.0.7
dbstream==0.1.28
defusedxml==0.8.0rc2
Django==5.1.4
django-allauth==65.3.1
django-axes==8.0.0
django-countries==7.6.1
django-crispy-forms==2.3
fonttools==4.55.3
google-api-core==2.24.0
google-api-python-client==2.160.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.22.1
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
libcst==1.5.1
lxml==5.3.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pyparsing==3.2.1
pypdf==5.1.0
pyphen==0.17.0
pypng==0.20220715.0
python-bidi==0.6.3
python-decouple==3.8
python-http-client==3.3.7
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.0
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==80.4.0
six==1.17.0
social-auth-core==4.5.4
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
stripe==12.1.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
wheel==0.45.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
